RF: Move serialization of genesis to separate file.
